reorganize tests files and fixed Path
