+++ client/components/QuestionsAndAnswers/AddQuestion.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import PropTypes from 'prop-types';
 import Button from './styles/SharedQA.style';
 import { QuestionModal } from './ModalForm';
 
@@ -15,6 +16,10 @@ const AddQuestion = ({ handleAddQuestion }) => {
       {open && (<QuestionModal setOpen={setOpen} handleAddQuestion={handleAddQuestion} />)}
     </div>
   );
+};
+
+AddQuestion.propTypes = {
+  handleAddQuestion: PropTypes.func.isRequired,
 };
 
 export default AddQuestion;
