+++ client/components/QuestionsAndAnswers/tests/MoreQuestions.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import MoreQuestions from './MoreQuestions';
+import MoreQuestions from '../MoreQuestions';
 
 it('render add a question + button correctly', () => {
   const { getByTestId } = render(<MoreQuestions />);
