+++ client/components/QuestionsAndAnswers/tests/QuestionItem.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import QuestionItem from './QuestionItem';
-import Data from './dummyData';
+import QuestionItem from '../QuestionItem';
+import Data from '../dummyData';
 
 it('should render QuestionItem component correctly', () => {
   const { getByTestId } = render(<QuestionItem question={Data.questions.results[0]} />);
