+++ client/components/QuestionsAndAnswers/tests/QuestionList.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import QuestionList from './QuestionList';
-import Data from './dummyData';
+import QuestionList from '../QuestionList';
+import Data from '../dummyData';
 
 it('should render QuestionList Component on the DOM', () => {
   const { getByTestId } = render(<QuestionList questions={Data.questions.results} />);
