+++ client/components/QuestionsAndAnswers/tests/SearchQuestion.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import SearchQuestion from './SearchQuestion';
+import SearchQuestion from '../SearchQuestion';
 
 it('should render a search bar', () => {
   const { getByTestId } = render(<SearchQuestion />);
